requirements.txt
streamlit
pandas
psycopg2
altair
python-dotenv
os